requirements.txt
# GST Council RAG System Dependencies

# Core ML/Embedding
sentence-transformers>=2.2.0
chromadb>=0.4.0

# LLM Client
openai>=1.0.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Web Interface
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0

